+++ src/main/java/org/prgrms/spring_week1/Order/repository/OrderRepository.java
@@ -1,0 +1,8 @@
+package org.prgrms.spring_week1.Order.repository;
+
+import org.prgrms.spring_week1.Order.model.Order;
+
+public interface OrderRepository {
+
+    void insert(Order order);
+}
