+++ src/main/java/org/prgrms/spring_week1/Voucher/model/Voucher.java
@@ -1,4 +1,4 @@
-package org.prgrms.spring_week1.models;
+package org.prgrms.spring_week1.Voucher.model;
 
 import java.util.UUID;
