+++ src/main/java/org/prgrms/spring_week1/Voucher/model/VoucherStatus.java
@@ -1,4 +1,4 @@
-package org.prgrms.spring_week1.models;
+package org.prgrms.spring_week1.Voucher.model;
 
 public enum VoucherStatus {
     VALID,
