--- src/main/java/org/prgrms/spring_week1/repositories/OrderRepository.java
@@ -1,8 +1,0 @@
-package org.prgrms.spring_week1.repositories;
-
-import org.prgrms.spring_week1.models.Order;
-
-public interface OrderRepository {
-
-    void insert(Order order);
-}
