--- src/main/resources/logback-backup.xml
@@ -1,42 +1,0 @@
-<configuration>
-
-  <conversionRule
-    conversionWord="clr"
-    converterClass="org.springframework.boot.logging.logback.ColorConverter"/>
-
-  <!-- 속성들 -->
-  <property name="CONSOLE_LOG_PATTERN"
-    value="%clr(%d{HH:mm:ss.SSS}){cyan} [%thread] %clr(%-5level) %logger{36} - %msg%n"/>
-  <property name="FILE_LOG_PATTERN"
-    value="%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n"/>
-
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-    <!-- encoders are assigned the type
-         ch.qos.logback.classic.encoder.PatternLayoutEncoder by default -->
-    <encoder>
-      <pattern>${CONSOLE_LOG_PATTERN}</pattern>
-    </encoder>
-  </appender>
-
-  <appender name="ROLLING_FILE" class="ch.qos.logback.core.rolling.RollingFileAppender">
-    <file>logs/access.log</file>  <!--        최신기록 -->
-    <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
-      <fileNamePattern>logs/access-%d{yyyy-MM-dd}.log</fileNamePattern> <!-- 날이지나면 -->
-    </rollingPolicy>
-    <encoder>
-      <pattern>${FILE_LOG_PATTERN}</pattern>
-    </encoder>
-  </appender>
-
-  <!--   file   -->
-  <logger name="org.prgrms.spring_week1" level="error">
-    <appender-ref ref="ROLLING_FILE"/>
-  </logger>
-
-  <!--  console  -->
-  <root level="debug">
-    <appender-ref ref="STDOUT"/>
-  </root>
-
-
-</configuration>
